+++ tablesalt/spec/tablesalt/array_index_spec.rb
@@ -7,6 +7,28 @@ RSpec.describe Tablesalt::ArrayIndex do
 
   it { is_expected.to delegate_method(:[]).to(:index) }
 
+  describe "#initialze" do
+    subject { array_index.array }
+
+    context "when the array is flat" do
+      it { is_expected.not_to equal array }
+      it { is_expected.to eq array }
+    end
+
+    context "when the array is nested" do
+      let(:nested_array) { Faker::Hipster.words }
+      let(:array) { [ nested_array ] }
+
+      it "deep_dups the array" do
+        expect(array_index.array).not_to equal array
+        expect(array_index.array).to eq array
+
+        expect(array_index.array[0]).not_to equal array[0]
+        expect(array_index.array[0]).to eq array[0]
+      end
+    end
+  end
+
   describe "#index" do
     subject { array_index.index }
 
@@ -40,6 +62,16 @@ RSpec.describe Tablesalt::ArrayIndex do
 
       it { is_expected.to eq repeated_item_index_before }
       it { is_expected.to eq array.index(requested_item) }
+    end
+  end
+
+  describe "#freeze" do
+    before { array_index.freeze }
+
+    it "freeze its attributes as well as itself" do
+      expect(array_index.array).to be_frozen
+      expect(array_index.index).to be_frozen
+      expect(array_index).to be_frozen
     end
   end
 end