+++ redis_hash/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.9.5
+
+*Release Date*: 5/20/2019
+
+- No changes
+
 ## v0.9.4
 
 *Release Date*: 5/20/2019